#2051 First Pass Checks

+++ config/linux.py
@@ -58,7 +58,8 @@ class LinuxConfig(AbstractConfig):
         except KeyError:
             logger.info("Config section not found. Backing up existing file (if any) and re-adding a section header")
             backup_filename = self.filename.parent / f'{appname}.ini.backup'
-            backup_filename.write_bytes(self.filename.read_bytes())
+            if self.filename.exists():
+                backup_filename.write_bytes(self.filename.read_bytes())
             self.config.add_section(self.SECTION)
 
         # Set 'outdir' if not specified or invalid
